Simplify engine component
